Moved Model and ModelData stubs to conftest for reusability
